Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/claim-fetch.https-expected.txt


PASS fetch() should be intercepted after the client is claimed.
